refactor: move Serialize interceptor to common directory
